+++ ecoville-back/src/main/java/senai/lab365/futurodev/ecoville_back/repository/SolicitacaoColetaRepository.java
@@ -5,7 +5,7 @@ import senai.lab365.futurodev.ecoville_back.entity.SolicitacaoColeta;
 import senai.lab365.futurodev.ecoville_back.enums.StatusColeta;
 import java.util.List;
 
-public interface SolicitacaoColetaRepository extends JpaRepository<SolicitacaoColeta, Long> {
-    List<SolicitacaoColeta> findByUsuarioResidencialId(Long usuarioId);
+public interface SolicitacaoColetaRepository extends JpaRepository<SolicitacaoColeta, Integer> {
+    List<SolicitacaoColeta> findByUsuarioResidencialId(Integer usuarioId);
     List<SolicitacaoColeta> findByStatus(StatusColeta status);
 }
